+++ R/morningstar.R
@@ -59,7 +59,7 @@
 #' from="2019-08-26",iuser = username, ipassword = password)
 #' }
 
-getPrice <- function(feed = "CME_NymexFutures_EOD",contract = "@CL0Z",
+getPrice <- function(feed = "CME_NymexFutures_EOD",contract = "@CL21Z",
                      from = "2020-09-01",iuser = "x@xyz.com", ipassword = "pass") {
   #mpurl <- "https://mp.morningstarcommodity.com/lds/feeds/CME_NymexFutures_EOD/ts?Symbol=@CL9Z"
   userpw <- paste0(iuser,":",ipassword)
@@ -91,25 +91,50 @@ getPrice <- function(feed = "CME_NymexFutures_EOD",contract = "@CL0Z",
     URL = httr::modify_url(url = "https://mp.morningstarcommodity.com",
                            path = paste0("/lds/feeds/",feed, "/ts?","cross_currencies=",x1,
                                          "&period=",x2,"&fromDateTime=",from))
-    }
+  }
+  if (feed %in% c("ERCOT_LmpsByResourceNodeAndElectricalBus")) {URL = httr::modify_url(url = "https://mp.morningstarcommodity.com",path = paste0("/lds/feeds/",feed, "/ts?","SettlementPoint=",contract,"&fromDateTime=",from))}
+  if (feed %in% c("PJM_Rt_Hourly_Lmp")) {URL = httr::modify_url(url = "https://mp.morningstarcommodity.com",path = paste0("/lds/feeds/",feed, "/ts?","pnodeid=",contract,"&fromDateTime=",from))}
+
   httr::handle_reset(URL)
   es <- httr::GET(url = URL,httr::authenticate(user = iuser,password = ipassword,type = "basic")) #,httr::progress())
-
-  if(length(httr::content(es))>0) {
-    if(length(httr::content(es) %>% purrr::flatten() %>% .$series %>% .$values %>% purrr::flatten()) > 0) {
-      out<-dplyr::tibble(date=as.character(lubridate::ymd(httr::content(es) %>%
-                                                         purrr::flatten() %>% purrr::flatten() %>%
-                                                         .$dates)) %>% lubridate::ymd(),
-                      value=as.numeric(httr::content(es) %>%
-                                         purrr::flatten() %>% purrr::flatten() %>%
-                                         .$values %>% .[[1]] %>%
-                                         purrr::flatten())) %>%
-        dplyr::mutate(value=ifelse(is.nan(value),NA,value)) } else {
-          out=dplyr::tibble(date=character(),value=numeric(),fwdmnt=numeric(),fwdyr=numeric())}} else {
-            out=dplyr::tibble(date=character(),value=numeric(),fwdmnt=numeric(),fwdyr=numeric())}
-  if (length(colnames(out))==2) {colnames(out)[2] <- contract}
+  es <- httr::content(es)
+  elecFeeds = "ERCOT|PJM"
+  # Non electricity feeds
+  if (!grepl(elecFeeds, feed) & length(es) > 0) {
+    if(length(es %>% purrr::flatten() %>% .$series %>% .$values %>% purrr::flatten()) > 0) {
+      out <-
+        dplyr::tibble(date = as.character(lubridate::ymd(es %>% purrr::flatten() %>% purrr::flatten() %>% .$dates)) %>% lubridate::ymd(),
+                      value = as.numeric(es %>% purrr::flatten() %>% purrr::flatten() %>% .$values %>% .[[1]] %>% purrr::flatten())) %>%
+        dplyr::mutate(value = ifelse(is.nan(value), NA, value)) } else {
+          out = dplyr::tibble(
+            date = character(),
+            value = numeric(),
+            fwdmnt = numeric(),
+            fwdyr = numeric()
+          )}}
+  # electricy feeds
+  if (grepl(elecFeeds, feed) & length(es) > 0) {
+    if(length(es %>% purrr::flatten() %>% .$series %>% .$values %>% purrr::flatten()) > 0) {
+      if (grepl("ERCOT", feed)) {tz <- "CST" ; x = 1}
+      if (grepl("PJM", feed)) {tz <- "EST" ; x = 5}
+      out <-
+        dplyr::tibble(date = as.POSIXct(sub("T","",es %>% purrr::flatten() %>% purrr::flatten() %>% .$date %>% unlist()),tz = tz),
+                      value = as.numeric(es %>% purrr::flatten() %>% purrr::flatten() %>% .$values %>% .[[x]] %>% purrr::flatten())) %>%
+        dplyr::mutate(value = ifelse(is.nan(value), NA, value)) } else {
+          out = dplyr::tibble(
+            date = character(),
+            value = numeric(),
+            fwdmnt = numeric(),
+            fwdyr = numeric()
+          )}}
+  if (length(colnames(out)) == 2) {colnames(out)[2] <- sub("@","",contract)}
   return(out)
 }
+
+#getPrice2(feed = "CME_NymexFutures_EOD",contract = "@CL21Z",from = "2020-09-01",iuser, ipassword )
+#getPrice2(feed = "ERCOT_LmpsByResourceNodeAndElectricalBus",contract = "HB_NORTH",from = "2020-09-01",iuser, ipassword )
+#getPrice2(feed = "PJM_Rt_Hourly_Lmp",contract = "51287",from = "2020-09-01",iuser, ipassword )
+
 
 #' \code{getPrices}
 #' @description